rudimentary backprog

+++ Cneuro/global.cpp
@@ -1,6 +1,6 @@
 #include "global.h"
 
-std::vector<std::vector<int>> connectionMatrix;
+std::vector<std::vector<float>> connectionMatrix;
 std::vector<Neuron> neurons; // Definition
 std::random_device rd;
 std::mt19937 gen(rd());
