+++ Cneuro/global.h
@@ -5,7 +5,7 @@
 #include "neuron.h" // Assuming Neuron is defined in Neuron.h
 #include <random>
 
-extern std::vector<std::vector<int>> connectionMatrix;
+extern std::vector<std::vector<float>> connectionMatrix;
 extern std::vector<Neuron> neurons; // Declaration
 extern std::random_device rd;
 extern std::mt19937 gen;
